feat: test base images with test applications
