+++ core-base-image-test/go.mod
@@ -1,0 +1,3 @@
+module base-image-test-service
+
+go 1.25.1
